refactoring JWT stuff, moving internal http functions
